Code Using LoaderManager
Code updated using LoaderManager and AsyncTaskLoader instead of
AsyncTask
